Add basic routing
Closes #27
