+++ data/ingredients/ingredients.json
@@ -1,0 +1,12 @@
+
+{ "ingredients": [
+    {"name":"flour"},
+    {"name":"sugar"},
+    {"name":"eggs"},
+    {"name":"butter"},
+    {"name":"milk"},
+    {"name":"vanilla"},
+    {"name":"baking powder"},
+    {"name":"salt"}
+  ]
+}
