+++ data/recipes/recipe-1.txt
@@ -1,0 +1,1 @@
+This is a test recipe.
