+++ public/pages/contact.html
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html lang='en'>
+<head>
+  <meta charset='UTF-8'>
+  <meta name='viewport' content='width=device-width, initial-scale=1.0'>
+  <title>Contact us!</title>
+</head>
+<body>
+  <header>
+    <h1>Contact us!</h1>
+  </header>
+  <main>
+    <p>Here is how you can contact us:</p>
+    <form action='/contact' method='post'>
+
+      <label for='name'>Name</label>
+      <input type='text' name='name' id='name'>
+
+      <label for='email'>Email</label>
+      <input type='email' name='email' id='email'>
+
+      <label for='message'>Message</label>
+      <textarea name='message' id='message'></textarea>
+
+      <input type='submit' value='Send'>
+    </form>
+  </main>
+  <footer>
+    <p>Thank you!</p>
+  </footer>
+</body>
+</html>
