+++ server/routes.js
@@ -1,0 +1,8 @@
+export { routes as default };
+
+const routes = {
+  '/': './public/ui/index.html',
+  '/recipes': './data/recipes/recipe-1.txt',
+  '/ingredients': './data/ingredients/ingredients.json',
+  '/contact': './public/pages/contact.html'
+};
